+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Added a new class on the Python framework: `InputValidation`.
+- New request: Delete a list of groups - `DELETE /agents/groups - Params: {"ids":["id_1",...,"id_n"]}`
 
 ### Fixed
 - When adding a new agent, now it checks that its name is different than manager's name.